chore: make the local test server more general
